Implement GetWallpaperUseCase
Add the ability to run the use case with a press of a button.
Also add the ability to view a simplified presenter output.

This will be handled with refresh or button events in the future.
The current implementation only serves for testing and as a proof
of concept for the Use case actually working.